+++ src/sql/dispensers.sql
@@ -1,7 +1,7 @@
 DROP TABLE IF EXISTS dispensers;
 CREATE TABLE dispensers (
-    tx_index       INTEGER UNSIGNED,
-    address_id     INTEGER UNSIGNED,
+    tx_index       BIGINT UNSIGNED,
+    address_id     BIGINT UNSIGNED,
     expiration     DATETIME,
     PRIMARY KEY(tx_index, address_id)
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8 COLLATE=utf8_unicode_ci;
