+++ src/sql/pubkeys.sql
@@ -1,6 +1,6 @@
 DROP TABLE IF EXISTS pubkeys;
 CREATE TABLE pubkeys (
-  address_id INTEGER UNSIGNED NOT NULL PRIMARY KEY,
+  address_id BIGINT UNSIGNED NOT NULL PRIMARY KEY,
   pubkey     VARCHAR(66) NOT NULL,
   FOREIGN KEY (address_id) REFERENCES index_addresses(id)
 );
